refactor: Quiz component
