--- src/app/admin/courses/[courseId]/module/[moduleId]/quiz/[quizQuestions]/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-type Props = {}
-
-const Quiz = (props: Props) => {
-    return <div>Quiz</div>
-}
-
-export default Quiz
